budget-service/requirements.txt
Django>=4.2
djangorestframework
kafka-python
